requirements.txt
albumentations
numpy
pillow
pyyaml
scipy
scikit-image
tqdm
torch
torchvision
torchaudio